Rename layers-config file
