+++ src/components/biodiversity-sidebar-card/biodiversity-sidebar-card.js
@@ -5,10 +5,10 @@ import metadataActions from 'redux_modules/metadata';
 import metadataService from 'services/metadata-service';
 import isEmpty from 'lodash/isEmpty';
 import Component from './biodiversity-sidebar-card-component';
-import { LAYERS_CATEGORIES } from 'constants/mol-layers-configs';
+import { LAYERS_CATEGORIES } from 'constants/layers-config';
 import { batchToggleLayers } from 'utils/layer-manager-utils';
 import mapStateToProps from './biodiversity-sidebar-card-selectors';
-import { biodiversityCategories } from 'constants/mol-layers-configs';
+import { biodiversityCategories } from 'constants/layers-config';
 import { useSelectLayersOnTabChange } from './biodiversity-sidebar-card-hooks';
 import { BIODIVERSITY_TABS_SLUGS } from 'constants/ui-params';
 const actions = {...metadataActions, ...urlActions};
