+++ src/components/landscape-sidebar/conservation-efforts-widget/conservation-efforts-widget.js
@@ -6,7 +6,7 @@ import conservationEffortsActions from 'redux_modules/conservation-efforts';
 import { addLayerAnalyticsEvent, removeLayerAnalyticsEvent } from 'actions/google-analytics-actions';
 import { layerManagerToggle } from 'utils/layer-manager-utils';
 import { batchToggleLayers } from 'utils/layer-manager-utils';
-import { layersConfig, LAYERS_CATEGORIES } from 'constants/mol-layers-configs';
+import { layersConfig, LAYERS_CATEGORIES } from 'constants/layers-config';
 import { COMMUNITY_AREAS_VECTOR_TILE_LAYER, GRID_CELLS_PROTECTED_AREAS_PERCENTAGE } from 'constants/layers-slugs';
 import { COMMUNITY_PROTECTED_AREAS_LAYER_GROUP } from 'constants/layers-groups';
 import { PROTECTED_AREAS_COLOR, COMMUNITY_AREAS_COLOR } from 'constants/protected-areas';
