+++ src/components/legend/legend-selectors.js
@@ -3,7 +3,7 @@ import { uniqBy } from 'lodash';
 import { getConfig } from 'utils/user-config-utils';
 import { LEGEND_FREE_LAYERS, LAND_HUMAN_PRESURES_LAYERS, MARINE_HUMAN_PRESURES_LAYERS, COMMUNITY_PROTECTED_AREAS_LAYER_GROUP } from 'constants/layers-groups';
 import { PLEDGES_LAYER, MARINE_AND_LAND_HUMAN_PRESSURES, COMMUNITY_AREAS_VECTOR_TILE_LAYER } from 'constants/layers-slugs';
-import { legendConfigs, DEFAULT_OPACITY } from 'constants/mol-layers-configs';
+import { legendConfigs, DEFAULT_OPACITY } from 'constants/layers-config';
 import { legendConfigs as humanPressureLegendConfigs, legendSingleRasterTitles } from 'constants/human-pressures';
 import { legendConfigs as WDPALegendConfigs } from 'constants/protected-areas';
 import { LEGEND_TUTORIAL, LEGEND_DRAG_TUTORIAL } from 'constants/tutorial';
