+++ src/components/priority-places-polygons-layer/priority-places-polygons-layer.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { loadModules } from 'esri-loader';
 import { PRIORITY_PLACES_POLYGONS, PRIORITY_POLYGONS_GRAPHIC_LAYER } from 'constants/layers-slugs';
-import { layersConfig } from 'constants/mol-layers-configs';
+import { layersConfig } from 'constants/layers-config';
 
 const PriorityPlacesPolygonsLayer = ({ view, selectedFeaturedMap, selectedTaxa }) => {
   const priorityPolygonsInitialState = {
